[Swift/C++ interop] Pass SDK to header generation
The header generation phase was failing to build because Swift couldn't
find the standard library modules. The standard library on Windows is
saved in the SDKROOT environment variable, so we can pass that. It would
be nice if the CMake Sysroot variable was set or we could do something a
little better for cross-compiling though.