[Swift/C++ interop] Find Swiftrt file
The Windows toolchains keep the Swift runtime in the SDK instead of in
the toolchain. The driver lists the runtime locations for the toolchain,
but doesn't list them for the SDK, so we don't really have a great way
to ask for the location. Instead, we translate the system processor to
the name of the architecture directory in the swift SDK and grab the
file directly.