Adding CMake version support ranges
Adding the supported version ranges to each project. The newer versions
of CMake have better support with some of the newer policies.
Specifically CMP0157 in CMake 3.29 improves the incremental rebuild
performance and hooks up the emissions of compile commands for Swift so
that folks with Sourcekit-LSP can have a better editing experience.