CMake4.0 Fixes: Only pass sdk to swift when set
CMake 4.0 does not specify CMAKE_OSX_SYSROOT unless it is set
explicitly. When this is the case, passing the `-sdk` flag with an empty
path confuses the compiler and results in build failures.

This patch fixes things to only pass the sdk to the Swift compiler when
the SDK is actually set.